Add selection to tasks closes #31
Add feature: make tasks dismissible

Add feature: close dismissible when selection is enabled

Add selectable widget

Add selection no notifications page

+++ lib/domain/repositories/queues_repository/queues_repository.dart
@@ -36,7 +36,7 @@ abstract class QueuesRepository {
 
   Future<void> completeTask({
     required int queueId,
-    double? expenses,
+    int? expenses,
   });
 
   Future<void> skipTask(int queueId);
