+++ lib/presentation/pages/add_progress_page.dart
@@ -4,7 +4,7 @@ import 'package:flutter/material.dart';
 import '../presentation.dart';
 
 class AddProgressPage extends StatefulWidget {
-  final void Function(double) submitExpenses;
+  final void Function(int) submitExpenses;
 
   const AddProgressPage({required this.submitExpenses, super.key});
 
@@ -45,7 +45,7 @@ class _AddProgressPageState extends State<AddProgressPage> {
             onPressed: !isValueValid
                 ? null
                 : () {
-                    widget.submitExpenses(double.parse(expenses));
+                    widget.submitExpenses(int.parse(expenses));
                     context.router.pop();
                   },
           ),
